Added TypeMap to Cartridge registry
