+++ std/cli/cli.ts
@@ -2,7 +2,6 @@ import { parse as parseFlags } from "../../deps/std/flags.ts";
 import { compile } from "../../lib/fart.ts";
 import { getFlags } from "./common.ts";
 import cartridges from "../carts/mod.ts";
-import typemaps from "../typemaps/mod.ts";
 
 export const cli = async (args: string[]) => {
   const flags = parseFlags(args);
@@ -15,11 +14,9 @@ export const cli = async (args: string[]) => {
     Deno.exit(0);
   }
   const content = await Deno.readTextFile(source);
-  const cartridge = cartridges.vendor(target);
-  if (cartridge === undefined) Deno.exit(1);
-  const targetLang = target?.split(".").shift();
-  const typemap = typemaps.vendor(targetLang);
-  if (typemap === undefined) Deno.exit(1);
+  const item = cartridges.vendor(target);
+  if (item === undefined) Deno.exit(1);
+  const { cartridge, typemap } = item;
   const code = await compile(content, {
     cartridge,
     typemap,
