+++ std/common.ts
@@ -12,3 +12,25 @@ export const convertFilenameToTargetFilename = (
     name: info.name,
   }).replace("\\", "/");
 };
+
+export enum Mime {
+  TypeScript = "application/typescript; charset=UTF-8",
+  JSON = "application/json; charset=UTF-8",
+  HTML = "text/html; charset=UTF-8",
+  CSS = "text/css; charset=UTF-8",
+  Plain = "text/plain; charset=UTF-8",
+  PNG = "image/png",
+}
+
+export const getMimeType = (pathname: string): Mime =>
+  pathname.endsWith(".css")
+    ? Mime.CSS
+    : pathname.endsWith(".html")
+    ? Mime.HTML
+    : pathname.endsWith(".json")
+    ? Mime.JSON
+    : pathname.endsWith(".ts")
+    ? Mime.TypeScript
+    : pathname.endsWith(".png")
+    ? Mime.PNG
+    : Mime.Plain;
