+++ std/server/common.ts
@@ -1,24 +1,4 @@
-export enum Mime {
-  TypeScript = "application/typescript; charset=UTF-8",
-  JSON = "application/json; charset=UTF-8",
-  HTML = "text/html; charset=UTF-8",
-  CSS = "text/css; charset=UTF-8",
-  Plain = "text/plain; charset=UTF-8",
-  PNG = "image/png",
-}
-
-export const getMimeType = (pathname: string): Mime =>
-  pathname.endsWith(".css")
-    ? Mime.CSS
-    : pathname.endsWith(".html")
-    ? Mime.HTML
-    : pathname.endsWith(".json")
-    ? Mime.JSON
-    : pathname.endsWith(".ts")
-    ? Mime.TypeScript
-    : pathname.endsWith(".png")
-    ? Mime.PNG
-    : Mime.Plain;
+import { Mime } from "../common.ts";
 
 export const makeError = (message: string, status = 400): Response =>
   new Response(JSON.stringify({ error: message }), {
